Expose byte index in tokenizer's position.
I need it to finish my implementation of Properties and Values syntax.

It works on bytes rather than CSS tokens, but it calls the consume a name
algorithm, for which I'm using cssparser.

However I need to know how much of the input was actually consumed afterwards to
continue the parse.