Refactor color parser to take a generic as output
This refactor allows the color parser to construct new color values
directly, in stead of constructing a predefined color value and then
forcing lib users to convert that type into whichever format they want
to use in the end.

Most common use case is that the color included in the lib,
cssparser::Color, is good for serialization, but not well suited for
doing calculations on.